First Matt commit.
Added Bluetooth (via BluetoothController)
Added PathHolder, which saves the points in a Path rather than just the
Path.
Changed the way gestureRecognizer works; works now on all roms.

All points are currently displayed as toasts rather then send to the
plotter for debugging reasons, but a full Byte encoding and decoding is
used.

And a lot more fixes and stuff I cant remember. :-)